moving layouts to app dir
