+++ app/account/create-account/_components/RegisterForm.tsx
@@ -8,7 +8,7 @@ import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
 
 import Button from '~/_components/button/Button';
 
-import Container from '~/layouts/Container';
+import Container from '~/_layouts/Container';
 import swell from '~/lib/SwellJS';
 import { notifyFailure, notifySuccess } from '~/utils/toastifies';
