+++ app/blog/page.tsx
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 
+import Container from '~/_layouts/Container';
 import keywords from '~/data/keywords.json';
-import Container from '~/layouts/Container';
 
 const { NEXT_PUBLIC_BASE_URL } = process.env;
