+++ app/privacy-policy/page.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 
 import keywords from '~/data/keywords.json';
-import Container from '~/layouts/Container';
+import Container from '~/_layouts/Container';
 
 const { NEXT_PUBLIC_BASE_URL } = process.env;
