+++ app/products/[slug]/page.tsx
@@ -5,7 +5,7 @@ import ProductSection from './_components/ProductSection';
 import RelatedProducts from './_components/RelatedProducts';
 
 import keywords from '~/data/keywords.json';
-import Container from '~/layouts/Container';
+import Container from '~/_layouts/Container';
 import Store from '~/lib/Store';
 
 interface ProductProp {
